requirements.txt
cryptography
python-dotenv
PyJWT
requests
